refactor: #666 pure and secure keystore.ts wip...
