+++ packages/core/src/utils/const/index.ts
@@ -1,6 +1,5 @@
 export * from './data';
 export * from './hdnode';
-export * from './keystore';
 export * from './mnemonic';
 export * from './secp256k1';
 export * from './transaction';
