--- packages/core/src/utils/const/keystore.ts
@@ -1,10 +1,0 @@
-/**
- * Keystore Scrypt params
- */
-const SCRYPT_PARAMS = {
-    N: 131072,
-    r: 8,
-    p: 1
-};
-
-export { SCRYPT_PARAMS };
